requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.8.0
python-dotenv==1.0.0
asyncpg==0.29.0
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 